+++ MeepMeep/src/main/java/club/wagnerrobotics/meepmeep/Visualizer.java
@@ -13,7 +13,7 @@ public class Visualizer {
                 .setConstraints(60, 60, Math.toRadians(180), Math.toRadians(180), 15)
                 .build();
 
-        myBot.runAction(myBot.getDrive().actionBuilder(new Pose2d(0, 0, 0))
+        myBot.runAction(myBot.getDrive().actionBuilder(new Pose2d(-60, 35, 0))
                 .lineToX(30)
                 .turn(Math.toRadians(90))
                 .lineToY(30)
@@ -24,7 +24,7 @@ public class Visualizer {
                 .turn(Math.toRadians(90))
                 .build());
 
-        meepMeep.setBackground(MeepMeep.Background.FIELD_POWERPLAY_OFFICIAL)
+        meepMeep.setBackground(MeepMeep.Background.FIELD_INTO_THE_DEEP_JUICE_DARK)
                 .setDarkMode(true)
                 .setBackgroundAlpha(0.95f)
                 .addEntity(myBot)