+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/wagner/mechanisms/Hanger.java
@@ -1,60 +1,32 @@
 package org.firstinspires.ftc.teamcode.wagner.mechanisms;
 
+import androidx.annotation.NonNull;
+
 import com.acmerobotics.dashboard.config.Config;
 import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.Gamepad;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 import com.qualcomm.robotcore.hardware.Servo;
 
-import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.firstinspires.ftc.teamcode.MecanumDrive;
 import org.firstinspires.ftc.teamcode.wagner.PartsMap;
 
 @Config
 public class Hanger implements Mechanism {
-    public DcMotor motor;
-    public static int OPEN_POSITION = 1000;
-    public static int CLOSED_POSITION = 0;
-    public static double power = 1.0;
-    public static boolean activation = false;
-    public static MechanismState state = MechanismState.UNEXTENDED;
+    public Servo hook;
+    public static double EXTENDED = 0.2;
+    public static double UNEXTENDED = 0;
 
     @Override
     public void init(HardwareMap hardwareMap) {
-        motor = hardwareMap.get(DcMotor.class, PartsMap.HANGER.toString());
-        motor.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-        motor.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        hook = hardwareMap.get(Servo.class, PartsMap.HANGER_HOOK.toString());
     }
 
-    public void internalGoTo(int position) {
-        motor.setTargetPosition(position);
-        motor.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-        motor.setPower(power);
-    }
+    public void extend() {hook.setPosition(EXTENDED);}
+    public void unextend() {hook.setPosition(UNEXTENDED);}
+    public void set(double pos) {hook.setPosition(pos);}
 
-    public MechanismState unextend() {
-        if(state == MechanismState.IN_USE) return state;
-        if(motor.isBusy()) return MechanismState.IN_USE;
-        activation = false;
-        internalGoTo(CLOSED_POSITION);
-        motor.setPower(0);
-        state = MechanismState.UNEXTENDED;
-        return state;
-    }
-
-    public MechanismState extend() {
-        if(state == MechanismState.IN_USE) return state;
-        if(motor.isBusy()) return MechanismState.IN_USE;
-        activation = true;
-        internalGoTo(OPEN_POSITION);
-        motor.setPower(0);
-        state = MechanismState.EXTENDED;
-        return state;
-    }
-
-    public MechanismState forceStop() {
-        activation = false;
-        motor.setPower(0);
-        return MechanismState.UNKNOWN;
+    @NonNull
+    public String toString() {
+        return "-- [Mechanism: Hanger] --\n" +
+                Mechanism.servoIfo(hook, "Hook");
     }
 }