+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/wagner/opmodes/NGTeleOp.java
@@ -2,11 +2,11 @@ package org.firstinspires.ftc.teamcode.wagner.opmodes;
 
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.config.Config;
+import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
 import com.acmerobotics.dashboard.telemetry.TelemetryPacket;
 import com.acmerobotics.roadrunner.PoseVelocity2d;
 import com.acmerobotics.roadrunner.Vector2d;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
 import org.firstinspires.ftc.teamcode.Drawing;
 import org.firstinspires.ftc.teamcode.MecanumDrive;
@@ -14,12 +14,12 @@ import org.firstinspires.ftc.teamcode.wagner.GlobalStorage;
 import org.firstinspires.ftc.teamcode.wagner.mechanisms.Arm;
 import org.firstinspires.ftc.teamcode.wagner.mechanisms.Claw;
 import org.firstinspires.ftc.teamcode.wagner.mechanisms.Hanger;
-import org.firstinspires.ftc.teamcode.wagner.mechanisms.MechanismState;
 
 @Config
 public class NGTeleOp extends LinearOpMode {
     @Override
     public void runOpMode() throws InterruptedException {
+        telemetry = new MultipleTelemetry(FtcDashboard.getInstance().getTelemetry(), telemetry);
         MecanumDrive drive = new MecanumDrive(hardwareMap, GlobalStorage.pose);
 
         Claw claw = new Claw();
@@ -60,32 +60,28 @@ public class NGTeleOp extends LinearOpMode {
             else if (gamepad2.dpad_down)
                 claw.down();
             else if (gamepad2.dpad_right)
-                claw.clawZero();
+                claw.middle();
 
             if (gamepad2.y) hanger.extend();
-            if (gamepad2.x) hanger.unextend();
-            if (gamepad2.right_stick_button) hanger.forceStop();
+            else hanger.unextend();
 
-            telemetry.addData("Hanger STATE", Hanger.state);
-            telemetry.addData("Hanger ACTIVATED", Hanger.activation);
-            telemetry.addData("Hanger MOTOR/Position", hanger.motor.getCurrentPosition());
-            telemetry.addData("Hanger MOTOR/Is Busy", hanger.motor.isBusy());
-            telemetry.addData("Hanger MOTOR/Power", hanger.motor.getPower());
-            telemetry.addData("Hanger MOTOR/Target Pos", hanger.motor.getTargetPosition());
-            telemetry.addData("Claw CLAW/Position", claw.claw.getPosition());
-            telemetry.addData("Claw WRIST/Position", claw.wrist.getPosition());
-            telemetry.addData("Arm ELBOW/Power", arm.elbow.getPower());
-            telemetry.addData("Arm SHOULDER/Power", arm.shoulder.getPower());
-            telemetry.addData("Arm SlideArm/LEFT Position", arm.left.getCurrentPosition());
-            telemetry.addData("Arm ArmArm/RIGHT Position", arm.right.getCurrentPosition());
-            telemetry.addLine("Gamepad 1: Drive");
-            telemetry.addLine("Gamepad 1 Right Trigger DOWN: Claw Open");
-            telemetry.addLine("Gamepad 1 Right Trigger Up: Claw Close");
-            telemetry.addLine("Gamepad 2 Right Stick Y: Arm Slides");
-            telemetry.addLine("Gamepad 2 Left Stick X: Elbow");
-            telemetry.addLine("Gamepad 2 Y: hanger.extend");
-            telemetry.addLine("Gamepad 2 X: hanger.unextend");
-            telemetry.addLine("Gamepad 2 Right Stick Press: hanger.forceStop");
+            telemetry.addLine("Gamepad 1");
+            telemetry.addLine("\tLeft & Right Stick: Drive");
+            telemetry.addLine();
+            telemetry.addLine("Gamepad 2");
+            telemetry.addLine("\tLeft Stick X: Elbow");
+            telemetry.addLine("\tLeft Stick Y: Shoulder");
+            telemetry.addLine("\tRight Stick Y: Arm Slides");
+            telemetry.addLine("\tY: Hanger Hook Extend (HOLD)");
+            telemetry.addLine("\tRight Trigger: Claw Close (HOLD)");
+            telemetry.addLine("\tD-Pad:");
+            telemetry.addLine("\t\tUp: Claw UP");
+            telemetry.addLine("\t\tDown: Claw DOWN");
+            telemetry.addLine("\t\tRight: Claw Middle");
+            telemetry.addLine();
+            telemetry.addLine(arm.toString());
+            telemetry.addLine(claw.toString());
+            telemetry.addLine(hanger.toString());
             telemetry.update();
 
             TelemetryPacket packet = new TelemetryPacket();