+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/wagner/opmodes/SillyTeleOpRegistrar.java
@@ -2,6 +2,7 @@ package org.firstinspires.ftc.teamcode.wagner.opmodes;
 
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.config.reflection.ReflectionConfig;
+import com.acmerobotics.dashboard.config.variable.CustomVariable;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.OpModeManager;
 import com.qualcomm.robotcore.eventloop.opmode.OpModeRegistrar;
@@ -17,7 +18,7 @@ public final class SillyTeleOpRegistrar {
         "808 Essentially Means Crashout",
         "Estoy Gay: The Return",
         "New Gen TeleOp",
-        ""
+        "Spicy Bucket lives on"
     };
     public SillyTeleOpRegistrar() {}
 
@@ -32,8 +33,10 @@ public final class SillyTeleOpRegistrar {
                     .setSource(OpModeMeta.Source.BLOCKLY)
                     .build();
         manager.register(meta, new NGTeleOp());
+        CustomVariable cv = ReflectionConfig.createVariableFromClass(NGTeleOp.class);
+        if(cv.size() == 0) return;
         FtcDashboard.getInstance().withConfigRoot(configRoot -> {
-            configRoot.putVariable(meta.name, ReflectionConfig.createVariableFromClass(NGTeleOp.class));
+            configRoot.putVariable(meta.name, cv);
         });
     }
 }