+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/wagner/opmodes/mechanismtests/TestHanger.java
@@ -6,27 +6,16 @@ import com.acmerobotics.roadrunner.Vector2d;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.teamcode.MecanumDrive;
-import org.firstinspires.ftc.teamcode.wagner.GlobalStorage;
 import org.firstinspires.ftc.teamcode.wagner.mechanisms.Hanger;
 
 public class TestHanger extends LinearOpMode {
     @Override
     public void runOpMode() throws InterruptedException {
-        MecanumDrive drive = new MecanumDrive(hardwareMap, new Pose2d(0,0,0));
         Hanger hanger = new Hanger();
         hanger.init(hardwareMap);
 
         waitForStart();
 
-        PoseVelocity2d input = new PoseVelocity2d(
-                new Vector2d(
-                        -gamepad1.left_stick_y,
-                        -gamepad1.left_stick_x
-                ),
-                drive.pose.heading.toDouble()
-        );
-
-        drive.setDrivePowers(input);
-        telemetry.update();
+        
     }
 }