+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/wagner/opmodes/mechanismtests/TestRegistrar.java
@@ -2,11 +2,14 @@ package org.firstinspires.ftc.teamcode.wagner.opmodes.mechanismtests;
 
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.config.reflection.ReflectionConfig;
+import com.acmerobotics.dashboard.config.variable.CustomVariable;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.OpModeManager;
 import com.qualcomm.robotcore.eventloop.opmode.OpModeRegistrar;
 
 import org.firstinspires.ftc.robotcore.internal.opmode.OpModeMeta;
+import org.firstinspires.ftc.teamcode.wagner.GlobalStorage;
+import org.firstinspires.ftc.teamcode.wagner.opmodes.NGTeleOp;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -29,17 +32,19 @@ public final class TestRegistrar {
         for(Map.Entry<String, LinearOpMode> set : tests.entrySet()) {
             OpModeMeta meta = new OpModeMeta.Builder()
                     .setName("Test " + set.getKey())
-                    .setFlavor(OpModeMeta.Flavor.AUTONOMOUS)
-                    .setGroup("Mechanism Tests")
-                    .setSource(OpModeMeta.Source.BLOCKLY)
+                    .setFlavor(GlobalStorage.testOpModeFlavor)
+                    .setGroup(GlobalStorage.testOpModeGroup)
+                    .setSource(GlobalStorage.testOpModeSource)
                     .build();
             manager.register(meta, set.getValue());
         }
         FtcDashboard.getInstance().withConfigRoot(configRoot -> {
             for(Map.Entry<String, LinearOpMode> set : tests.entrySet()) {
+                CustomVariable cv = ReflectionConfig.createVariableFromClass(set.getValue().getClass());
+                if(cv.size() == 0) return;
                 configRoot.putVariable(
                         set.getValue().getClass().getSimpleName(),
-                        ReflectionConfig.createVariableFromClass(set.getValue().getClass())
+                        cv
                 );
             }
         });